requirements.txt
openai>=1.0.0
astropy>=5.3.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
